requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
pyodbc==5.2.0
fpdf==1.7.2
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
email-validator==2.1.0 